tests-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file tests-requirements.txt tests-requirements.in
#
astroid==2.0.4            # via pylint
atomicwrites==1.2.0       # via pytest
attrs==18.1.0             # via pytest
coverage==5.0a1           # via pytest-cov
isort==4.3.4              # via pylint
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via pylint
more-itertools==4.3.0     # via pytest
pg8000==1.12.3            # via testing-postgresql
pluggy==0.7.1             # via pytest, tox
py==1.6.0                 # via pytest, tox
pylint==2.1.1
pytest-cov==2.5.1
pytest-django==3.4.2
pytest==3.7.3             # via pytest-cov, pytest-django
python-box==3.2.0
six==1.11.0               # via astroid, more-itertools, pg8000, pytest, tox
testing-postgresql==1.3.0
testing.common.database==2.0.3  # via testing-postgresql
tox==3.2.1
virtualenv==16.0.0        # via tox
wrapt==1.10.11            # via astroid
